OSINT Reconnaissance Framework/requirements.txt
Flask==3.0.0
python-whois==0.8.0
dnspython==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==2.1.0